Fix bug where placeholder text could contain empty char + normal text and then be falsely identified as only containing empty chars. This would ultimately lead to the entire text being replaced with one empty char.
