+++ frontend/src/plate/components.ts
@@ -12,7 +12,7 @@ import { ListItem, OrderedList, UnorderedList } from '@app/plate/components/list
 import { Maltekst } from '@app/plate/components/maltekst/maltekst';
 import { PageBreak } from '@app/plate/components/page-break';
 import { Paragraph } from '@app/plate/components/paragraph';
-import { Placeholder } from '@app/plate/components/placeholder';
+import { Placeholder } from '@app/plate/components/placeholder/placeholder';
 import { TableCellElement } from '@app/plate/components/plate-ui/table-cell-element';
 import { TableElement } from '@app/plate/components/plate-ui/table-element';
 import { TableRowElement } from '@app/plate/components/plate-ui/table-row-element';
